Fixes some stying, moves loader into the app, gets rid of header
